Moved stream api function to resources
